features-ML/requirements.txt
pandas
numpy
scikit-learn
datetime
flask
python-dotenv
pyjwt
flask_cors 
mysql-connector-python
xgboost